requirements.txt
streamlit

pillow